Extend functionality of message queue
A use case we ran into was needing to be able to see if there are
currently any messages related to the event in the queue, in order to
see if the user has interrupted the bot, or vice-versa.  Along with
this, we also wanted the capability to cancel the requests that were
found.  Both of these were added as `peek` and `cancelAll` which both
require the current event.