chore(meta): add integrity hashes to yarn.lock files
The hashes are added automatically when you run the latest Yarn version. So good to get this merged so it won't show up as added lines when people run `$ yarn` in the repo. It doesn't change any functionality. This is merely to reduce noise.

More info: https://github.com/yarnpkg/yarn/pull/5042